refactor: integrate 5 headers into 1 header
