+++ src/components/Header.js
@@ -1,7 +1,16 @@
-import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-import { devices } from '../utils/StyleUtils';
+import { Link, useNavigate, useLocation } from 'react-router-dom';
+import styled from 'styled-components/macro';
+import { signOut } from 'firebase/auth';
+import { useContext, useEffect, useState } from 'react';
+import {
+  Search, FavoriteBorder, Favorite, ContentCopy, Edit,
+} from '@styled-icons/material-rounded';
+import PropTypes from 'prop-types';
 import logoImage from '../images/CookYourWay_logo_v1.png';
+import { auth } from '../firestore/index';
+import { devices } from '../utils/StyleUtils';
+import AuthContext from './AuthContext';
+import chefImage from '../images/chef.png';
 
 const Background = styled.div`
   width: 100%;
@@ -39,6 +48,7 @@ const RightDiv = styled(LeftDiv)`
 const ImgLink = styled(Link)`
   width: calc(64*100vw/1920);
   height: calc(64*100vw/1920);
+  display: flex;
   @media ${devices.Tablet} and (orientation:portrait) {
     width: calc(180*100vw/1920);
     height: calc(180*100vw/1920);
@@ -48,10 +58,6 @@ const ImgLink = styled(Link)`
 const TitleLink = styled(Link)`
   text-decoration: none;
   color: inherit;
-`;
-
-const StyledLink = styled(Link)`
-  text-decoration: none;
 `;
 
 const Img = styled.img`
@@ -70,10 +76,190 @@ const Title = styled.div`
   font-size: calc(48*100vw/1920);
   @media ${devices.Tablet} and (orientation:portrait) {
     width: calc(1000*100vw/1920);
-    height: calc(180*100vw/1920);
+    height: calc(200*100vw/1920);
     font-size: calc(120*100vw/1920);
     padding-top: calc(20*100vw/1920);
   }
+`;
+
+const CreateButton = styled.button`
+  width: calc(200*100vw/1920);
+  height: calc(64*100vw/1920);
+  background-color: #EB811F;
+  cursor: pointer;
+  border: 0;
+  border-radius: calc(15*100vw/1920);
+  font-size: calc(28*100vw/1920);
+  color:#FDFDFC;
+  &:hover{
+  background-color:#fa8921;
+  }
+  @media ${devices.Tablet} and (orientation:portrait) {
+    width: calc(300*100vw/1920);
+    height: calc(150*100vw/1920);
+    font-size: calc(70*100vw/1920);
+  }
+`;
+
+const ButtonLink = styled(Link)`
+  width: calc(200*100vw/1920);
+  height: calc(64*100vw/1920);
+  display: flex;
+  text-decoration: none;
+  @media ${devices.Tablet} and (orientation:portrait) {
+    width: calc(300*100vw/1920);
+    height: calc(150*100vw/1920);
+  }
+`;
+
+const SignOutButton = styled.button`
+  width: calc(200*100vw/1920);
+  height: calc(64*100vw/1920);
+  background: transparent;
+  cursor: pointer;
+  border-radius: calc(15*100vw/1920);
+  font-size: calc(28*100vw/1920);
+  color: #2B2A29;
+  border: 0;
+  @media ${devices.Tablet} and (orientation:portrait) {
+    width: calc(300*100vw/1920);
+    height: calc(150*100vw/1920);
+    font-size: calc(70*100vw/1920);
+  }
+`;
+
+const MemberPhoto = styled.img`
+  width: calc(64*100vw/1920);
+  height: calc(64*100vw/1920);
+  cursor: pointer;
+  border-radius: 50%;
+  display: ${(props) => (props.active ? 'flex' : 'none')};
+  justify-content: center;
+  align-items: center;
+  box-shadow: 0px 0px 3px #e0e0e0;
+  object-fit: cover;
+  @media ${devices.Tablet} and (orientation:portrait) {
+    width: calc(160*100vw/1920);
+    height: calc(160*100vw/1920);
+  }
+`;
+
+const MemberWord = styled.div`
+  width: calc(64*100vw/1920);
+  height: calc(64*100vw/1920);
+  background-color: #FDFDFC;
+  cursor: pointer;
+  border-radius: 50%;
+  font-size: calc(28*100vw/1920);
+  color: #2B2A29;
+  border: 0;
+  display: ${(props) => (props.active ? 'flex' : 'none')};
+  justify-content: center;
+  align-items: center;
+  box-shadow: 0px 0px 3px #e0e0e0;
+  @media ${devices.Tablet} and (orientation:portrait) {
+    width: calc(160*100vw/1920);
+    height: calc(160*100vw/1920);
+    font-size: calc(70*100vw/1920);
+  }
+`;
+
+const DropdownContentDiv = styled.div`
+  display: none;
+  position: absolute;
+  left: calc(-160*100vw/1920);
+  background-color: #fff;
+  min-width: calc(160*100vw/1920);
+  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
+  z-index: 1;
+  transition: .5s;
+  border-radius: calc(15*100vw/1920);
+  &:hover{
+    background-color: #f0f0f0;
+  }
+`;
+
+const DropdownDiv = styled.div`
+  position: relative;
+  display: inline - block;
+  &:hover ${DropdownContentDiv} {display: block;};
+`;
+
+const Padding = styled.div`
+  height: calc(116*100vw/1920);
+  @media ${devices.Tablet} and (orientation:portrait) {
+    height: calc(320*100vw/1920);
+  }
+`;
+
+const SearchWrap = styled.div`
+  display: inline-block;
+  position: relative;
+  height: calc(116*100vw/1920);
+  @media ${devices.Tablet} and (orientation:portrait) {
+    height: calc(320*100vw/1920);
+  }
+`;
+
+const SearchForm = styled.form``;
+
+const SearchInputImg = styled.div`
+  height: calc(62*100vw/1920);
+  width: calc(62*100vw/1920);
+  display: inline-block;
+  border: none;
+  position: absolute;
+  top: calc(30*100vw/1920);
+  right: 0;
+  z-index: 2;
+  cursor: pointer;
+  transition: opacity 0.4s ease;
+  opacity: 0.4;
+  &:hover {
+  opacity: 0.8;
+  }
+  @media ${devices.Tablet} and (orientation:portrait) {
+    top: calc(90*100vw/1920);
+    height: calc(150*100vw/1920);
+    width: calc(150*100vw/1920);
+  }
+`;
+
+const SearchInput = styled.input`
+  height: calc(80*100vw/1920);
+  display: inline-block;
+  border: none;
+  outline: none;
+  padding-right: calc(80*100vw/1920);
+  width: 0;
+  position: absolute;
+  top: calc(24*100vw/1920);
+  right: 0;
+  background: none;
+  z-index: ${(props) => (props.show ? 3 : 1)};
+  transition: width 0.4s cubic-bezier(0, 0.795, 0, 1);
+  cursor: pointer;
+  &:hover ~ ${SearchInputImg}{
+    opacity: 0.8;
+  }
+  &:focus:hover {
+  border-bottom: calc(2.55*100vw/1920) solid #343638;
+  }
+  &:focus {
+  font-size: calc(36*100vw/1920);
+  color: #343638;
+  width: calc(800*100vw/1920);
+  z-index: 1;
+  border-bottom: calc(2.55*100vw/1920) solid #34363850;
+  cursor: text;
+  }
+  @media ${devices.Tablet} and (orientation:portrait) {
+    display: none;
+  }
+`;
+
+const StyledLink = styled(Link)`
+  text-decoration: none;
 `;
 
 const DropBtn = styled.button`
@@ -101,32 +287,289 @@ const DropBtn = styled.button`
   }
 `;
 
-const Padding = styled.div`
-  height: calc(116*100vw/1920);
+const ToolTipText = styled.span`
+  visibility: hidden;
+  width: calc(440*100vw/1920);
+  background-color: #26262590;
+  font-size: calc(26*100vw/1920);
+  color: #fff;
+  text-align: center;
+  border-radius: 6px;
+  padding: calc(20*100vw/1920); 0;
+  position: absolute;
+  z-index: 1;
+  bottom: -230%;
+  left: -100%;
+  margin-left: calc(-250*100vw/1920);;
+  opacity: 0;
+  transition: opacity 1s;
+`;
+
+const Icon = styled.span`
+  position: relative;
+  cursor: pointer;
+  font-size: calc(40*100vw/1920);
+  display: flex;
+  align-items: end;
+  color:  ${(props) => ((props.colorCode))};
+  &:hover {
+    color: ${(props) => ((props.selected))};
+  }
+  & > svg{
+    width: calc(60*100vw/1920);
+    height: calc(60*100vw/1920);
+  }
+  & > svg:hover {
+    color: ${(props) => ((props.selected))};
+  }
+  &:hover > ${ToolTipText} {
+    visibility: visible;
+    opacity: 1;
+  }
   @media ${devices.Tablet} and (orientation:portrait) {
-    height: calc(320*100vw/1920);
+    & > svg{
+    width: calc(160*100vw/1920);
+    height: calc(160*100vw/1920);
+  }
   }
 `;
 
-function Header() {
+const Button = styled(Link)`
+  display: flex;
+  text-decoration: none;
+`;
+
+function SignOut() {
+  const navigate = useNavigate();
+  function handleSignOut() {
+    signOut(auth).then(() => {
+      navigate('/login', { replace: true });
+      // Sign-out successful.
+    });
+  }
+
+  return (
+    <SignOutButton type="button" onClick={() => handleSignOut()}>登出</SignOutButton>
+  );
+}
+
+function SearchRecipe() {
+  const [inputShown, setInputShown] = useState(true);
+  const [searchName, setSearchName] = useState('');
+  const navigate = useNavigate();
+  return (
+    <SearchWrap>
+      <SearchForm action="" autocomplete="on">
+        <SearchInput
+          show={inputShown}
+          name="search"
+          type="search"
+          placeholder="來找些料理吧！"
+          onKeyPress={(e) => {
+            if (e.key === 'Enter') {
+              navigate({ pathname: '/search_recipe', search: `?q=${e.target.value}` });
+            }
+          }}
+          onChange={(e) => { setSearchName(e.target.value); }}
+          onBlur={() => {
+            setInputShown(false);
+          }}
+        />
+        <SearchInputImg>
+          <Search onClick={() => {
+            if (searchName) {
+              navigate({ pathname: '/search_recipe', search: `?q=${searchName}` });
+            } else {
+              setInputShown(true);
+            }
+          }}
+          />
+        </SearchInputImg>
+      </SearchForm>
+    </SearchWrap>
+  );
+}
+
+function CreateRecipe() {
+  return (
+    <ButtonLink to="/modify_recipe">
+      <CreateButton>寫食譜</CreateButton>
+    </ButtonLink>
+  );
+}
+
+function LogOut() {
+  const userInfo = useContext(AuthContext);
+  const [userInitial, setUserInitial] = useState('');
+  const [showMemberWord, setShowMemberWord] = useState(true);
+  const [showMemberPhoto, setShowMemberPhoto] = useState(false);
+
+  useEffect(() => {
+    if (userInfo) {
+      const userName = userInfo.name.split('');
+      const initial = userName.shift();
+      setUserInitial(initial.toUpperCase());
+    }
+  }, [userInfo]);
+
+  return (
+    <DropdownDiv>
+      <MemberWord
+        active={showMemberWord}
+        onMouseLeave={() => { setShowMemberWord(true); setShowMemberPhoto(false); }}
+        onMouseEnter={() => { setShowMemberPhoto(true); setShowMemberWord(false); }}
+      >
+        {userInitial}
+      </MemberWord>
+      <MemberPhoto
+        active={showMemberPhoto}
+        onMouseLeave={() => { setShowMemberWord(true); setShowMemberPhoto(false); }}
+        onMouseEnter={() => { setShowMemberPhoto(true); setShowMemberWord(false); }}
+        src={chefImage}
+        alt="memberImage"
+      />
+      <DropdownContentDiv>
+        <SignOut />
+      </DropdownContentDiv>
+    </DropdownDiv>
+  );
+}
+
+function LogIn() {
+  return (
+    <StyledLink to="/login">
+      <DropBtn>登入</DropBtn>
+    </StyledLink>
+  );
+}
+
+function CollectRecipe({
+  addToFavorites, removeFromFavorites, myFavorites, currentRecipeId,
+}) {
+  return (
+
+    myFavorites.includes(currentRecipeId)
+      ? (
+        <Icon colorCode="#BE0028" selected="#dd1b44" onClick={removeFromFavorites}>
+          <Favorite />
+          <ToolTipText>
+            點擊取消收藏
+            <br />
+            嗚嗚～我的心碎了一地
+          </ToolTipText>
+        </Icon>
+      )
+      : (
+        <Icon colorCode="#808080" selected="#b2b0b0" onClick={addToFavorites}>
+          <FavoriteBorder />
+          <ToolTipText>喜歡我嗎？趕緊加入收藏吧！</ToolTipText>
+        </Icon>
+      )
+  );
+}
+
+function CopyOrModifyRecipe({ userId, authorId }) {
+  const location = useLocation();
+  return (
+    <Button to={`/modify_recipe?id=${location.search.split('=')[1]}`}>
+      {userId === authorId
+        ? (
+          <Icon colorCode="#808080" selected="#b2b0b0">
+            <Edit />
+            <ToolTipText>有些新想法嗎？ 來編輯食譜吧！</ToolTipText>
+          </Icon>
+        )
+        : (
+          <Icon colorCode="#808080" selected="#b2b0b0">
+            <ContentCopy />
+            <ToolTipText>
+              複製食譜！
+              <br />
+              將天馬行空的點子寫進食譜吧！
+            </ToolTipText>
+          </Icon>
+        )}
+    </Button>
+  );
+}
+
+function Header({
+  authorId, userId, addToFavorites, removeFromFavorites, myFavorites, currentRecipeId,
+}) {
+  const location = useLocation();
+  const isHomePage = location.pathname === '/home';
+  const isHomeOrSearchPage = location.pathname === '/home' || location.pathname === '/search_recipe';
+  const isLandingPage = location.pathname === '/';
+  const isReadRecipePage = location.pathname === '/read_recipe';
+
   return (
     <>
       <Background>
         <LeftDiv>
-          <ImgLink to="/">
+          <ImgLink to="/home">
             <Img src={logoImage} alt="logoImage" />
           </ImgLink>
-          <TitleLink to="/"><Title>Cook Your Way</Title></TitleLink>
+          <TitleLink to="/home">
+            <Title>Cook Your Way</Title>
+          </TitleLink>
         </LeftDiv>
         <RightDiv>
-          <StyledLink to="/login">
-            <DropBtn>登入</DropBtn>
-          </StyledLink>
+          {isHomePage ? <SearchRecipe /> : ''}
+          {isHomeOrSearchPage ? <CreateRecipe /> : ''}
+          {isReadRecipePage
+            ? (
+              <CollectRecipe
+                addToFavorites={addToFavorites}
+                myFavorites={myFavorites}
+                removeFromFavorites={removeFromFavorites}
+                currentRecipeId={currentRecipeId}
+              />
+            )
+            : ''}
+          {isReadRecipePage
+            ? (
+              <CopyOrModifyRecipe
+                authorId={authorId}
+                userId={userId}
+              />
+            )
+            : ''}
+          {isLandingPage ? <LogIn /> : <LogOut />}
         </RightDiv>
       </Background>
       <Padding />
     </>
   );
 }
+
+Header.propTypes = {
+  authorId: PropTypes.string,
+  userId: PropTypes.string,
+  addToFavorites: PropTypes.func,
+  removeFromFavorites: PropTypes.func,
+  myFavorites: PropTypes.arrayOf(PropTypes.string),
+  currentRecipeId: PropTypes.string,
+};
+
+Header.defaultProps = {
+  authorId: '',
+  userId: '',
+  addToFavorites: () => { },
+  removeFromFavorites: () => { },
+  myFavorites: [],
+  currentRecipeId: '',
+};
+
+CollectRecipe.propTypes = {
+  addToFavorites: PropTypes.func.isRequired,
+  removeFromFavorites: PropTypes.func.isRequired,
+  myFavorites: PropTypes.arrayOf(PropTypes.string).isRequired,
+  currentRecipeId: PropTypes.string.isRequired,
+};
+
+CopyOrModifyRecipe.propTypes = {
+  authorId: PropTypes.string.isRequired,
+  userId: PropTypes.string.isRequired,
+};
 
 export default Header;
